fixing bugs - simple programs work

+++ src/single_interpreters/boxing.hpp
@@ -5,14 +5,14 @@
 #pragma once
 #include <cstdint>
 
-constexpr uint32_t box(uint32_t value) {
+constexpr int64_t box(int64_t value) {
     return (value << 1) | 1;
 }
 
-constexpr uint32_t unbox(uint32_t value) {
+constexpr int64_t unbox(int64_t value) {
     return value >> 1;
 }
 
-constexpr bool is_boxed(uint32_t value) {
+constexpr bool is_boxed(int64_t value) {
     return value & 1;
 }
